requirements.txt
requests>=2.31.0
rich>=13.7.0
prompt_toolkit>=3.0.43
pygments>=2.17.0
gitpython>=3.1.40
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
aiohttp>=3.9.1
pillow>=10.1.0
pydantic>=2.5.0
watchdog>=3.0.0
pymongo>=4.6.0
dnspython>=2.4.0
anthropic>=0.25.0
